requirements.txt
# ============================================
# 小红书爬虫项目依赖文件
# Python版本要求: 3.13.2
# ============================================

# ============================================
# Web框架
# ============================================
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
swagger-ui-bundle==0.0.9

# ============================================
# 异步HTTP客户端
# ============================================
httpx==0.27.2
aiohttp==3.10.11

# ============================================
# 数据库相关
# ============================================
# 异步MySQL驱动
aiomysql==0.2.0
# SQLAlchemy异步支持
sqlalchemy==2.0.36
# 数据库连接池
asyncpg==0.30.0

# ============================================
# 对象存储
# ============================================
minio==7.2.6

# ============================================
# 异步任务队列
# ============================================
celery==5.4.0
redis==5.2.0
aioredis==2.0.1

# ============================================
# 爬虫相关
# ============================================
# 异步爬虫框架
playwright==1.48.0
selenium==4.27.1
# HTML解析
beautifulsoup4==4.12.3
lxml==5.3.0
# 请求处理
requests==2.32.3

# ============================================
# 环境配置
# ============================================
python-dotenv==1.0.1

# ============================================
# 日志和监控
# ============================================
loguru==0.7.2

# ============================================
# 工具类
# ============================================
# 时间处理
python-dateutil==2.9.0
# 数据验证
validators==0.34.0
# 加密解密
cryptography==43.0.1

# ============================================
# 开发工具（可选）
# ============================================
# pytest==8.3.3
# pytest-asyncio==0.24.0
# black==24.8.0
# flake8==7.1.1
# mypy==1.11.2
